pyproject.toml
[tool.poetry]
name = "airbnb-clone-backend2"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "airbnb_clone_backend2"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
pillow = "^9.4.0"
djangorestframework = "^3.14.0"
strawberry-graphql = {extras = ["debug-server"], version = "^0.168.2"}
strawberry-graphql-django = "^0.9.4"
pyjwt = "^2.6.0"
django-environ = "^0.10.0"
django-cors-headers = "^3.14.0"
requests = "^2.28.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
